pdPalette: start work on palette exporting in popular formats
First up is Adobe's traditional swatch format (.aco).  pdStream needed
to be expanded to support writing various bits in big-endian format, and
since this was something I wanted for a future effort to build my own
PSD writer, this was a good time to tackle it.

The goal here is ultimately to support export for all formats that we
currently import, so there's plenty left to do.  :/