pdPalette: add support for writing modern ASE files (Adobe Swatch Exchange)
And with this commit, PD can now export all the palette formats it's
capable of importing!  This is far more than I ever thought we'd
accomplish, and this class has probably ended up as the most versatile
palette converter on the planet.  (Not sure how useful that is, but hey
- it's there.)

Now I need to figure out the best way to expose these features to users.
A File > Export menu is probably due for implementation, with the option
to export the active image's color scheme to any supported palette file.
Similarly, there is possibly also wisdom in having palette formats
exposed right inside the program's main Load/Save dialogs (as it could
be convenient to load a palette file, then render it as a nice grid of
colors, just to see what it looks like).  That would also provide a
"roundabout" way to convert palettes from one format to another.

The main problem at present is that there's really nowhere in the
program where a dedicated palette converter fits.  I'll keep mulling
this over while I work on some other items.

(Also, ASE export support was a blocker for #249.  Now that it's
available, we can combine all useful "Retro" palettes into a single ASE
file, then ship that with the program, rather than trying to maintain a
list of which palette files in a PD subfolder are relevant to a Retro
tool and which are not.)