pdButton: implement new owner-drawn mode
This is the joy of running on a custom widget set: if I need a new UI
feature, it takes all of several minutes to add the required options to
a PD control, vs several days/weeks of trying to coerce new behavior out
of Windows common controls.

In particular, I need custom rendering for a new "options" button on the
color selection panel.  The available space is very tight, too tight to
use a normal text caption, but I can easily custom draw a "..." caption.
This gives us a touch-accessible way to let the user modify the color
selection panel's layout and options, as opposed to requiring something
like a right-click popup menu.