Right toolbar: further reduce size requirements
As much as I like mirrored left/right padding (it looks so pretty!),
it's more important to free up space for the primary canvas.  The
right-side toolbar now requires even less room than it previously did,
and it will also default to a smaller size when run for the first time.

Also, a settings button now lives on the color panel.  It doesn't
actually *do* anything yet, but it will shortly!