Right-side toolbar: new approach to panel resizing
Thank you to @Kroc for the suggestion to use titlebars for resizing
instead of dead space between panels.

This new solution is quite a bit more complicated, and I'm not entirely
sure I've accounted for everything, but it seems to be working
reasonably well...?  The greatest complications arise from handling the
mouse drag events in the titlebars themselves (which are standalone
windows), then reporting those events back to our parent window, which
then moves the titlebars around accordingly.  Synchronizing the mouse
positions during these changes is non-trivial, because mouse coordinates
are reported per-hWnd but the hWnds in question are moving as a result
of the mouse actions... but by translating everything to screen
coordinates as I go, I believe I've worked out the most egregious kinks.
I still find occasional "blips" where the windows don't appear to
synchronize perfectly to mouse drags, but I can't reproduce the problem
reliably, so I'll need to investigate further.

Also, I need to create a proper mouse cursor solution that represents
the new "clickable and draggable" behavior, to make this whole effort
actually discoverable.

While here, I tried to improve rendering performance on various
right-side panels.  There's still work to be done but I've eked out a
few extra fps across various rendering operations.  The right-side
toolbar is a messy one as many of its settings have to be synchronized
to image changes (e.g. the navigator box, layer settings, etc), so it is
redrawn with greater frequency than other parts of the main UI.  As
such, performance wins here are more significant than performance wins
elsewhere in the UI.