Tooltip manager: controls can request "instant" tooltip display...
...instead of waiting for a MouseHover event.

This fixes #253, I believe - but please let me know if I missed
anything!

At present, only the "Color Variants", "Color Wheel", and "Palette"
controls make use of this behavior by default.  The main color panel's
"History" control also uses this, but it is not baked into the history
control by default.  (That control is completely owner-drawn, so it's up
to specific implementations to decide how tooltips are rendered.)