pdFFT: start working toward making FFTs usable
A number of advanced image processing techniques rely on FFTs, and while
I've had a basic FFT implementation in PD for awhile, no tools are tied
it, primarily because it's not yet fast enough for serious use.

I'm not sure if I can rectify this (given VB's inherent limitations),
but I'm willing to try - and if nothing else, having code that's ready
to "plug-in" to a working FFT implementation will make it easier to
switch to a 3rd-party library if it comes to that.

In this commit, I've added a bunch of functions necessary to move image
data to/from FFT representation, which allows me to formally confirm
that our FFT implementation works.  Yay?

I'll work on optimization amidst other work, but for now, let's play it
safe and say that features tied to FFTs remain in the "maybe someday"
realm of possibilities.