Rulers: finalize notch strategy (for px mode, anyway)
Canvas rulers are now fully functional.  Yay!

At present, I have only implemented pixel measurements.  Inch and cm are
still to-do, and my current plan is to synchronize ruler units with the
measurement unit dropdown in the canvas statusbar.  Inches in particular
will require a different notch strategy (as 1/4 and 1/8 inch
measurements are preferable to e.g. 1/10), but I'm saving this for a
future commit.

For now, enjoy pixel-accurate measurements (I hope - I've tested across
a wide variety of images; more testing is always welcome).  In
particular, I've tried hard to ensure that ruler and status bar canvas
coordinates line up exactly - or at least, as exactly as is physically
possible, given the current zoom value.

The rulers render in real-time, and there is no measurable performance
impact when they are turned on.  So feel free to use them or not use
them at your leisure; it makes no difference for viewport performance.

(FYI - #42 is nearly solved as of this commit, but I'll wait to close it
until inch and cm measurements are implemented.)