Suspend specialized ExifTool post-processing for PNG files
Because FreeImage has any number of deficiencies when processing PNG
files, we used ExifTool to fill in some of the gaps.  The problem with
that is that ExifTool runs asynchronously, so it may finish processing
metadata long after the image has already been parsed and displayed
on-screen - so when we update pixel bits, there's a disconcerting sudden
change of the image's on-screen appearance.

But we don't need to play these games anymore.  Our internal PNG decoder
can process everything ExifTool does and more, so I'm removing all of
the specialized ExifTool-based post-processing steps for gAMA, cHRM,
palette handling, and other PNG elements.