Faster effect previews when a selection is active
Effect previews in general are due for an overhaul (as required for #252, for example), but in the meantime, I may as well clean-up a few items.

For example, selection effect rendering can be improved by caching a relevant copy of the selection mask pre-sized to match the preview area.  This saves a meaningful amount of time when the image and/or selection is large, and it'll be very helpful for generating a real-time preview gallery.