pdRuler: proper fix for vertical font rendering across Windows versions
Turns out Windows 7 has totally different vertical font rendering behavior, too - thanks to contributor shishi for reporting this to me, and thanks to Microsoft for making this so much harder than it needs to be!  /s

I now have to make various run-time font rendering decisions based on the current OS - and on Windows 7, I have to do this even though the OS supplies a vertically optimized font.  I'm honestly not 100% sure what is going on under the hood here, because Win 7 and Win 10 should render similarly... but that's life as a Windows developer, I guess.  (sigh)